Extract core HMR client logic to shared runtime
This extracts the core HMR client logic, abstracting away hooks for platform-specific features like the websocket transport and page reload.

Test Plan: CI, existing HMR tests
